Refactored the HomeViewModel to adapt to MapViewModel. These changes include: 1. Not giving a map as argument to DisplayItinerary in HomeScreen.kt which was silly. Now just have to give the itinerary and the navigation to know where to go. 2. Removed a function from the ItineraryRepository and several from the HomeViewModel 3. Moved all in a home package that contains useful functions to display the itinerary
