+++ app/src/main/java/com/example/triptracker/view/home/HomeScreen.kt
@@ -1,30 +1,20 @@
-package com.example.triptracker.view
+package com.example.triptracker.view.home
 
 import android.util.Log
-import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.CornerSize
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material.icons.outlined.Star
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.MaterialTheme
@@ -38,7 +28,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.platform.testTag
@@ -48,13 +37,11 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import coil.compose.AsyncImage
 import com.example.triptracker.R
 import com.example.triptracker.model.itinerary.Itinerary
+import com.example.triptracker.view.Navigation
+import com.example.triptracker.view.NavigationBar
 import com.example.triptracker.view.theme.md_theme_grey
-import com.example.triptracker.view.theme.md_theme_light_black
-import com.example.triptracker.view.theme.md_theme_light_onPrimary
-import com.example.triptracker.view.theme.md_theme_orange
 import com.example.triptracker.viewmodel.HomeViewModel
 
 /**
@@ -104,105 +91,12 @@ fun HomeScreen(navigation: Navigation, homeViewModel: HomeViewModel = viewModel(
                 contentPadding = PaddingValues(16.dp)) {
                   items(itineraries) { itinerary ->
                     Log.d("ItineraryToDisplay", "Displaying itinerary: $itinerary")
-                    DisplayItinerary(
-                        itinerary = itinerary,
-                        navigation = navigation,
-                        homeViewModel.pinNamesMap.value ?: emptyMap())
+                    DisplayItinerary(itinerary = itinerary, navigation = navigation)
                   }
                 }
           }
         }
       }
-}
-
-/**
- * Displays an itinerary in the list of itineraries
- *
- * @param itinerary: Itinerary object to display
- * @param navigation: Navigation object to use for navigation
- * @param pinNamesMap: Map of itinerary ID to list of pin names
- */
-@Composable
-fun DisplayItinerary(
-    itinerary: Itinerary,
-    navigation: Navigation,
-    pinNamesMap: Map<String, List<String>>
-) {
-  // Number of additional itineraries not displayed
-  val pinNames = pinNamesMap[itinerary.id]
-  val pinListString = pinNames?.let { convertPinListToString(it) }
-  Box(
-      modifier =
-          Modifier.fillMaxWidth()
-              .padding(vertical = 10.dp)
-              .background(color = md_theme_light_black, shape = RoundedCornerShape(35.dp))
-              .clickable { // For now, when you click on an itinerary, it brings you to the map
-                // to view the itinerary, however it doesn't also switch on the navbar
-                // TODO : when changing Top Level Destination, the navbar should be updated to
-                // highlight the correct tab.
-                val map = navigation.getTopLevelDestinations()[1]
-                navigation.navigateTo(map)
-                // TODO : Would call DisplayItineraryInMap or sth similar later on
-              }) {
-        Column(modifier = Modifier.fillMaxWidth().padding(25.dp)) {
-          Row(modifier = Modifier.fillMaxWidth()) {
-            // change the image to the user's profile picture
-            AsyncImage(
-                model =
-                    "https://blueprintdigital.com/wp-content/uploads/2014/11/stock-photo-happy-man-giving-okay-sign-portrait-on-white-background-141327337.jpg",
-                contentDescription = "User Avatar",
-                modifier = Modifier.size(20.dp).clip(CircleShape))
-            Spacer(modifier = Modifier.width(15.dp))
-            Text(
-                text = itinerary.username,
-                fontFamily = FontFamily(Font(R.font.montserrat_regular)),
-                fontWeight = FontWeight.Normal,
-                fontSize = 14.sp,
-                color = md_theme_grey)
-            Spacer(modifier = Modifier.width(120.dp))
-            Icon(
-                imageVector = Icons.Outlined.Star,
-                contentDescription = "Star",
-                Modifier.size(20.dp))
-          }
-          Spacer(modifier = Modifier.height(5.dp))
-          Log.d("ItineraryRoute", itinerary.route.toString())
-          Text(
-              text = itinerary.title,
-              fontFamily = FontFamily(Font(R.font.montserrat_regular)),
-              fontWeight = FontWeight.Bold,
-              fontSize = 24.sp,
-              color = md_theme_light_onPrimary)
-          Text(
-              text = "${itinerary.flameCount}🔥",
-              color = md_theme_orange, // This is the orange color
-              fontFamily = FontFamily(Font(R.font.montserrat_regular)),
-              fontSize = 14.sp)
-          Spacer(modifier = Modifier.height(20.dp))
-          Spacer(modifier = Modifier.heightIn(6.dp).weight(1f))
-          Text(
-              text = "$pinListString",
-              fontSize = 14.sp,
-              fontFamily = FontFamily(Font(R.font.montserrat_medium)),
-              color = md_theme_grey)
-        }
-      }
-}
-
-/**
- * Helper function to convert a list of pin names to a string
- *
- * @param pinList: List of pin names
- * @return String representation of the list of pin names
- */
-fun convertPinListToString(pinList: List<String>): String {
-  return if (pinList.size <= 3) {
-    pinList.joinToString(", ")
-  } else {
-    val displayedPins = pinList.take(3).joinToString(", ")
-    val remainingCount = pinList.size - 3
-    "$displayedPins, and $remainingCount more"
-  }
 }
 
 /**
@@ -311,8 +205,8 @@ fun SearchBar(
             onItineraryClick = { id ->
               // Hard coded for the moment
               // TODO: Navigate to the itinerary details screen (2nd screen Figma)
-              val map = navigation.getTopLevelDestinations()[1].route
-              navigation.navController.navigate(map)
+              val map = navigation.getTopLevelDestinations()[1]
+              navigation.navigateTo(map)
             })
       }
     }
