+++ app/src/main/java/com/example/triptracker/viewmodel/HomeViewModel.kt
@@ -1,14 +1,15 @@
 package com.example.triptracker.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
 import androidx.lifecycle.switchMap
+import androidx.lifecycle.viewModelScope
 import com.example.triptracker.model.itinerary.Itinerary
 import com.example.triptracker.model.itinerary.ItineraryList
 import com.example.triptracker.model.repository.ItineraryRepository
+import kotlinx.coroutines.launch
 
 /**
  * ViewModel for the Home Screen. Fetches all itineraries from the repository stores them in a
@@ -32,31 +33,13 @@ class HomeViewModel : ViewModel() {
 
   // Fetch all itineraries from the repository on initialization
   init {
-    fetchItineraries()
+    viewModelScope.launch { fetchItineraries() }
   }
 
   /** Fetches all itineraries from the repository and stores them in the itineraryList LiveData */
   private fun fetchItineraries() {
-    itineraryInstance.itineraryList =
-        repository.getAllItineraries(callback = { itineraryList -> fetchPinNames(itineraryList) })
+    itineraryInstance.itineraryList = repository.getAllItineraries()
     _itineraryList.value = itineraryInstance.itineraryList
-  }
-
-  /**
-   * Fetches the names of all pinned places for each itinerary in the list
-   *
-   * @param itineraries List of itineraries
-   */
-  private fun fetchPinNames(itineraries: List<Itinerary>) {
-    val pinNamesMap = mutableMapOf<String, List<String>>()
-
-    for (itinerary in itineraries) {
-      Log.d("TypeOfPin", itinerary.pinnedPlaces.javaClass.simpleName)
-      val pinNames = repository.getPinNames(itinerary)
-      Log.d("pinNamesMap", "Fetched pin names for itinerary: $pinNames")
-      pinNamesMap[itinerary.id] = pinNames
-    }
-    _pinNamesMap.postValue(pinNamesMap)
   }
 
   // Filtered Itinerary list LiveData based on search query
