+++ app/src/main/java/com/example/triptracker/viewmodel/MapViewModel.kt
@@ -1,14 +1,11 @@
 package com.example.triptracker.viewmodel
 
-import android.util.Log
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.triptracker.model.geocoder.NominatimApi
-import com.example.triptracker.model.itinerary.Itinerary
 import com.example.triptracker.model.itinerary.ItineraryList
-import com.example.triptracker.model.location.Location
 import com.example.triptracker.model.repository.ItineraryRepository
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.LatLngBounds
@@ -40,7 +37,6 @@ class MapViewModel : ViewModel() {
   val selectedPolylineState = mutableStateOf<SelectedPolyline?>(null)
 
   init {
-    updateAllItineraries()
     viewModelScope.launch { getAllItineraries() }
   }
 
@@ -57,7 +53,7 @@ class MapViewModel : ViewModel() {
 
   /** Get all itineraries from the database and update the pathList */
   private fun getAllItineraries() {
-    repository.getAllItineraries { itineraries -> _pathList.postValue(ItineraryList(itineraries)) }
+    _pathList.postValue(ItineraryList(repository.getAllItineraries()))
   }
 
   /**
@@ -90,7 +86,8 @@ class MapViewModel : ViewModel() {
    * This function is used to update all the itineraries in the database with random routes. This is
    * temporary code
    */
-  // TODO: Remove this function after the database is populated with real data
+  // TODO: REMOVE this function after the database is populated with real data
+  /*
   private fun updateAllItineraries() {
     repository.getAllItineraries { itineraries ->
       var i = 0
@@ -117,6 +114,8 @@ class MapViewModel : ViewModel() {
       }
     }
   }
+
+   */
 
   // TODO: Remove this function after the database is populated with real data
   private fun generateRandomCoordinates(): LatLng {
